requirements.txt
#celery==2.5.1
#celery-tasktree==0.3.2
#django-celery==2.2.4
django-extensions==0.9
django-multidb-router==0.5
django-nose==1.1
django-mysql-pool==0.2
#gunicorn==0.15.0
metlog-py==0.9.8
metlog-cef==0.2
metlog-raven==0.3
raven==2.0.6
PyMySQL==0.5


#-e git://github.com/andymckay/django-radagast@4ea83aa714787a0cc352d39aaed27473139fc9db#egg=django-radagast
-e git://github.com/clelland/django-mysql-pymysql.git@acd2b46a885537ab9aa8e777e86a1fe86775f9ea#egg=django-mysql-pymysql
-e git://github.com/jbalogh/jingo.git@b4ecca1066bcd1eac87358b4c82902442d4eedc4#egg=jingo
-e git://github.com/washort/gelato.models.git#egg=gelato.models
-e git://github.com/washort/django-waffle.git@19fe01f48008efc02e357554d49cc3b6e0b1c854#egg=django-waffle
-e git://github.com/fwenzel/django-mozilla-product-details.git@36ef06539d6b34c4f345fd0d3e16937d0db9a752#egg=django-mozilla-product-details
-e git://github.com/jsocol/jingo-minify.git@d1b668ad7c0423fe8cc639df36d433da5c168cac#egg=jingo-minify
